requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
transformers==4.36.2
torch==2.1.2
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
plotly==5.18.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
